+++ src/main/java/org/ppijerman/ppijidentitybackend/server/service/security/dao/authentication/PpijIdAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package org.ppijerman.ppijidentitybackend.server.service.security.dao;
+package org.ppijerman.ppijidentitybackend.server.service.security.dao.authentication;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
